dev_requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile dev_requirements.in --output-file dev_requirements.txt
-e file:.
    # via -r dev_requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
authlib==1.3.1
    # via flickr-photos-api
build==1.2.1
    # via -r dev_requirements.in
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
coverage==7.5.3
    # via pytest-cov
cryptography==42.0.8
    # via authlib
docutils==0.21.2
    # via readme-renderer
flickr-url-parser==1.9.0
    # via flickr-photos-api
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   flickr-photos-api
    #   flickr-url-parser
hyperlink==21.0.0
    # via flickr-url-parser
idna==3.7
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   yarl
importlib-metadata==7.1.0
    # via twine
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.1
    # via keyring
keyring==25.2.1
    # via twine
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.2.0
    # via
    #   jaraco-classes
    #   jaraco-functools
multidict==6.0.5
    # via yarl
mypy==1.10.0
    # via -r dev_requirements.in
mypy-extensions==1.0.0
    # via mypy
nh3==0.2.17
    # via readme-renderer
packaging==24.0
    # via
    #   build
    #   pytest
pkginfo==1.11.0
    # via twine
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via -r dev_requirements.in
pydantic-core==2.18.4
    # via pydantic
pygments==2.18.0
    # via
    #   readme-renderer
    #   rich
pyproject-hooks==1.1.0
    # via build
pytest==8.2.1
    # via pytest-cov
pytest-cov==5.0.0
    # via -r dev_requirements.in
pyyaml==6.0.1
    # via vcrpy
readme-renderer==43.0
    # via twine
requests==2.32.3
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.7.1
    # via twine
ruff==0.4.8
    # via -r dev_requirements.in
silver-nitrate==1.1.1
    # via flickr-photos-api
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==8.3.0
    # via flickr-photos-api
twine==5.1.0
    # via -r dev_requirements.in
typing-extensions==4.12.1
    # via
    #   mypy
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   requests
    #   twine
vcrpy==6.0.1
    # via -r dev_requirements.in
wrapt==1.16.0
    # via vcrpy
yarl==1.9.4
    # via vcrpy
zipp==3.19.1
    # via importlib-metadata
